+++ modules/apps/commerce/commerce-pricing-web/src/main/java/com/liferay/commerce/pricing/web/internal/constants/CommercePriceListScreenNavigationConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.pricing.web.internal.servlet.taglib.ui.constants;
+package com.liferay.commerce.pricing.web.internal.constants;
 
 /**
  * @author Alessio Antonio Rendina
