+++ modules/apps/commerce/commerce-pricing-web/src/main/resources/META-INF/resources/init.jsp
@@ -55,6 +55,9 @@ page import="com.liferay.commerce.pricing.exception.NoSuchPricingClassException"
 page import="com.liferay.commerce.pricing.model.CommercePriceModifier" %><%@
 page import="com.liferay.commerce.pricing.model.CommercePricingClass" %><%@
 page import="com.liferay.commerce.pricing.type.CommercePriceModifierType" %><%@
+page import="com.liferay.commerce.pricing.web.internal.constants.CommerceDiscountScreenNavigationConstants" %><%@
+page import="com.liferay.commerce.pricing.web.internal.constants.CommercePriceListScreenNavigationConstants" %><%@
+page import="com.liferay.commerce.pricing.web.internal.constants.CommercePricingClassScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.pricing.web.internal.constants.CommercePricingFDSNames" %><%@
 page import="com.liferay.commerce.pricing.web.internal.display.context.AddedAllCommerceDiscountRuleDisplayContext" %><%@
 page import="com.liferay.commerce.pricing.web.internal.display.context.AddedAnyCommerceDiscountRuleDisplayContext" %><%@
@@ -73,9 +76,6 @@ page import="com.liferay.commerce.pricing.web.internal.display.context.CommerceP
 page import="com.liferay.commerce.pricing.web.internal.display.context.CommercePricingClassDisplayContext" %><%@
 page import="com.liferay.commerce.pricing.web.internal.display.context.CommercePricingClassPriceListDisplayContext" %><%@
 page import="com.liferay.commerce.pricing.web.internal.display.context.CommerceTierCommercePriceEntryDisplayContext" %><%@
-page import="com.liferay.commerce.pricing.web.internal.servlet.taglib.ui.constants.CommerceDiscountScreenNavigationConstants" %><%@
-page import="com.liferay.commerce.pricing.web.internal.servlet.taglib.ui.constants.CommercePriceListScreenNavigationConstants" %><%@
-page import="com.liferay.commerce.pricing.web.internal.servlet.taglib.ui.constants.CommercePricingClassScreenNavigationConstants" %><%@
 page import="com.liferay.commerce.product.exception.NoSuchCatalogException" %><%@
 page import="com.liferay.commerce.product.model.CPDefinition" %><%@
 page import="com.liferay.commerce.product.model.CPInstance" %><%@
